pyproject.toml
[project]
name = "nSocCSP"
version = "0.3.0"
description = "Network Monitoring and Security Audit"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "networkx>=3.4.2",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "python-nmap>=0.7.1",
    "scapy>=2.6.1",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.45.0",
]
